Improves the code on the AddDataViewController to smooth things, and implemnents the method on dataStore to save and fetch the gender

+++ bodymass/Protocols/HeightSelectorDelegate.swift
@@ -9,5 +9,5 @@
 import Foundation
 
 protocol HeightSelectorDelegate: class {
-  func heightChanged(value: Float)
+  func heightChanged(value: Double)
 }
